[Global] - Config general font and split User Dropdown Menu away from Main Layout
